requirements.txt
absl-py==0.9.0
alabaster==0.7.12
apex==0.1
argon2-cffi==20.1.0
ascii-graph==1.5.1
ase==3.21.1
attrs==19.3.0
audioread==2.1.8
Babel==2.8.0
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==3.1.5
blis==0.4.1
boto3==1.14.45
botocore==1.17.45
brotlipy==0.7.0
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
codecov==2.1.8
conda==4.8.3
conda-build==3.18.11
conda-package-handling==1.7.0
coverage==5.2.1
cryptography==2.9.2
cxxfilt==0.2.2
cycler==0.10.0
cymem==2.0.3
Cython==0.28.4
dataclasses==0.8
DataProperty==0.50.0
decorator==4.4.2
defusedxml==0.6.0
DLLogger @ git+https://github.com/NVIDIA/dllogger@26a0f8f1958de2c0c460925ff6102a4d2486d6cc
docutils==0.15.2
easydict==1.9
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
Flask==1.1.2
future==0.18.2
glob2==0.7
googledrivedownloader==0.4
grpcio==1.31.0
h5py==2.10.0
html2text==2020.1.16
hypothesis==4.50.8
idna==2.9
imageio==2.9.0
imagesize==1.2.0
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1593446433964/work
inflect==4.1.0
iniconfig==1.0.1
ipdb==0.13.3
ipykernel==5.3.4
ipython @ file:///tmp/build/80754af9/ipython_1593447367857/work
ipython-genutils==0.2.0
isodate==0.6.0
itsdangerous==1.1.0
jedi @ file:///tmp/build/80754af9/jedi_1592841914522/work
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
json5==0.9.5
jsonschema==3.0.2
jupyter-client==6.1.6
jupyter-core==4.6.3
jupyter-tensorboard==0.2.0
jupyterlab==1.2.14
jupyterlab-server==1.2.0
jupytext==1.5.2
kiwisolver==1.2.0
libarchive-c==2.9
librosa==0.6.3
llvmlite==0.28.0
lmdb==0.99
Markdown==3.2.2
MarkupSafe==1.1.1
maskrcnn-benchmark @ file:///opt/pytorch/examples/maskrcnn/pytorch
matplotlib==3.3.1
mbstrdecoder==1.0.0
mccabe==0.6.1
mistune==0.8.4
mlperf-compliance==0.0.10
mock==4.0.2
more-itertools==8.4.0
msgfy==0.1.0
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.0
nltk==3.5
notebook==6.1.3
numba==0.43.1
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1596233718326/work
nvidia-dali-cuda110==0.24.0
onnx @ file:///opt/pytorch/pytorch/third_party/onnx
onnxruntime==1.4.0
opencv-python==3.4.1.15
packaging==20.4
pandas==0.24.2
pandocfilters==1.4.2
parso==0.7.0
pathvalidate==2.3.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
Pillow-SIMD @ file:///tmp/pillow-simd
pkginfo==1.5.0.1
plac @ file:///tmp/build/80754af9/plac_1594261902054/work
pluggy==0.13.1
preshed==3.0.2
progressbar==2.5
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.13.0
psutil==5.7.0
ptyprocess==0.6.0
py==1.9.0
pybind11==2.5.0
pycocotools @ git+https://github.com/nvidia/cocoapi.git@6ac4a93058202603f36fd1ce47228e7d81119e5a#subdirectory=PythonAPI
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.20
pydot==1.4.1
pyflakes==2.1.1
Pygments==2.6.1
pyldpc==0.7.9
pynvml==8.0.4
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1594392929924/work
pyparsing==2.4.7
pyprof @ file:///opt/pytorch/pyprof
pyrsistent==0.16.0
PySocks==1.7.1
pytablewriter==0.47.0
pytest==6.0.1
pytest-cov==2.10.1
pytest-pythonpath==0.7.3
python-dateutil==2.8.1
python-hostlist==1.20
python-louvain==0.15
python-nvd3==0.15.0
python-slugify==4.0.1
pytorch-transformers==1.1.0
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.2
rdflib==5.0.0
regex==2020.7.14
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
resampy==0.2.2
revtok @ git+git://github.com/jekbradbury/revtok.git@f1998b72a941d1e5f9578a66dc1c20b01913caab
ruamel-yaml==0.15.87
# Editable install with no version control (rvae==0.1)
-e /workspace/SPUD_Analysis-of-manifold-structure-in-Head_Direction-data
s3transfer==0.3.3
sacrebleu==1.2.10
sacremoses==0.0.35
scikit-image==0.15.0
scikit-learn==0.23.2
scipy @ file:///tmp/build/80754af9/scipy_1592930497347/work
Send2Trash==1.5.0
sentencepiece==0.1.91
six==1.15.0
snowballstemmer==2.0.0
SoundFile==0.10.3.post1
soupsieve==2.0.1
sox==1.4.0
spacy @ file:///tmp/build/80754af9/spacy_1594303279343/work
Sphinx==3.2.1
sphinx-rtd-theme==0.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
srsly==1.0.2
SSD @ file:///opt/pytorch/examples/ssd
subword-nmt @ git+git://github.com/rsennrich/subword-nmt.git@48ba99e657591c329e0003f0c6e32e493fa959ef
tabledata==1.1.3
tabulate==0.8.7
tensorboard @ file:///nvidia/opt/tensorboard_install/tensorboard-1.15.0%2Bnv-py3-none-any.whl
tensorboard-plugin-dlprof @ file:///nvidia/opt/tensorboard_install/tensorboard_plugin_dlprof-0.6-py3-none-any.whl
tensorrt==7.1.3.4
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
thinc @ file:///tmp/build/80754af9/thinc_1594251955397/work
threadpoolctl==2.1.0
toml==0.10.1
torch==1.7.1+cu110
torch-cluster==1.5.8
torch-geometric==1.6.3
torch-scatter==2.0.5
torch-sparse==0.6.8
torch-spline-conv==1.2.0
torchaudio==0.7.2
torchtext @ file:///opt/pytorch/text
torchvision==0.8.2+cu110
tornado==6.0.4
tqdm==4.31.1
traitlets==4.3.3
typepy==1.1.1
typing==3.7.4.3
typing-extensions==3.7.4.2
Unidecode==1.1.1
urllib3==1.25.9
wasabi==0.6.0
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.10.11
yacs==0.1.8
zipp==3.1.0
